# docs/api/openapi.yml
components:
  schemas:
...
    DecommissionOut:
      properties:
        node_id:
          title: Node Id
          type: string
        noop:
          title: Noop
          type: boolean
        status:
          title: Status
          type: string
      required:
      - node_id
      - status
      - noop
      title: DecommissionOut
      type: object
...
    GrantResultOut:
      properties:
        harbor_id:
          title: Harbor Id
          type: string
        noop:
          title: Noop
          type: boolean
        user_id:
          title: User Id
          type: string
      required:
      - user_id
      - harbor_id
      - noop
      title: GrantResultOut
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HarborAdminOut:
      properties:
        harbor_id:
          title: Harbor Id
          type: string
        lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Lat
        lng:
          anyOf:
          - type: number
          - type: 'null'
          title: Lng
        name:
          title: Name
          type: string
      required:
      - harbor_id
      - name
      title: HarborAdminOut
      type: object
    HarborCreate:
      properties:
        harbor_id:
          maxLength: 64
          minLength: 1
          title: Harbor Id
          type: string
        lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Lat
        lng:
          anyOf:
          - type: number
          - type: 'null'
          title: Lng
        name:
          maxLength: 128
          minLength: 1
          title: Name
          type: string
      required:
      - harbor_id
      - name
      title: HarborCreate
      type: object
    HarborGrant:
      properties:
        harbor_id:
          maxLength: 64
          minLength: 1
          title: Harbor Id
          type: string
      required:
      - harbor_id
      title: HarborGrant
      type: object
    HarborGrantOut:
      properties:
        harbor_id:
          title: Harbor Id
          type: string
        role:
          title: Role
          type: string
      required:
      - harbor_id
      - role
      title: HarborGrantOut
      type: object
    HarborOut:
      properties:
        harbor_id:
          examples:
          - harbor-saltsjobaden
          title: Harbor Id
          type: string
        name:
          examples:
          - Saltsjöbaden Marina
          title: Name
          type: string
      required:
      - harbor_id
      - name
      title: HarborOut
      type: object
    HarborPatch:
      properties:
        lat:
          anyOf:
          - type: number
          - type: 'null'
          title: Lat
        lng:
          anyOf:
          - type: number
          - type: 'null'
          title: Lng
        name:
          anyOf:
          - maxLength: 128
            minLength: 1
            type: string
          - type: 'null'
          title: Name
      title: HarborPatch
      type: object
...
    NodeHealthOut:
      properties:
        adopted_at:
          examples:
          - '2026-05-03T14:00:00Z'
          format: date-time
          title: Adopted At
          type: string
        battery_pct:
          anyOf:
          - examples: &id016
            - 87
            type: integer
          - type: 'null'
          examples: *id016
          title: Battery Pct
        berth_id:
          examples:
          - berth-001
          title: Berth Id
          type: string
        gateway_id:
          examples:
          - gw-dock-a
          title: Gateway Id
          type: string
        health:
          enum:
          - online
          - stale
          - offline
          - decommissioned
          title: Health
          type: string
        last_seen:
          anyOf:
          - examples: &id017
            - '2026-05-03T14:30:00Z'
            format: date-time
            type: string
          - type: 'null'
          examples: *id017
          title: Last Seen
        mesh_unicast_addr:
          examples:
          - '0x0042'
          title: Mesh Unicast Addr
          type: string
        node_id:
          examples:
          - node-012
          title: Node Id
          type: string
        serial_number:
          examples:
          - DP-N-000123
          title: Serial Number
          type: string
      required:
      - node_id
      - serial_number
      - berth_id
      - gateway_id
      - mesh_unicast_addr
      - adopted_at
      - health
      title: NodeHealthOut
      type: object
...
    UserAdminOut:
      properties:
        boat_club:
          anyOf:
          - type: string
          - type: 'null'
          title: Boat Club
        email:
          title: Email
          type: string
        firstname:
          title: Firstname
          type: string
        lastname:
          title: Lastname
          type: string
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        role:
          examples:
          - harbormaster
          title: Role
          type: string
        user_id:
          title: User Id
          type: string
      required:
      - user_id
      - email
      - firstname
      - lastname
      - role
      title: UserAdminOut
      type: object
    UserCreatedOut:
      properties:
        email:
          title: Email
          type: string
        role:
          title: Role
          type: string
        user_id:
          title: User Id
          type: string
      required:
      - user_id
      - email
      - role
      title: UserCreatedOut
      type: object
...
    UserPatchOut:
      properties:
        email:
          title: Email
          type: string
        role:
          title: Role
          type: string
        user_id:
          title: User Id
          type: string
      required:
      - user_id
      - email
      - role
      title: UserPatchOut
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    app__routers__admin__users__UserCreate:
      properties:
        boat_club:
          anyOf:
          - maxLength: 128
            type: string
          - type: 'null'
          title: Boat Club
        email:
          format: email
          title: Email
          type: string
        firstname:
          maxLength: 64
          minLength: 1
          title: Firstname
          type: string
        lastname:
          maxLength: 64
          minLength: 1
          title: Lastname
          type: string
        password:
          format: password
          maxLength: 128
          minLength: 8
          title: Password
          type: string
          writeOnly: true
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
        role:
          default: boat_owner
          enum:
          - harbormaster
          - boat_owner
          title: Role
          type: string
      required:
      - email
      - password
      - firstname
      - lastname
      title: UserCreate
      type: object
    app__routers__admin__users__UserPatch:
      properties:
        boat_club:
          anyOf:
          - maxLength: 128
            type: string
          - type: 'null'
          title: Boat Club
        firstname:
          anyOf:
          - maxLength: 64
            minLength: 1
            type: string
          - type: 'null'
          title: Firstname
        lastname:
          anyOf:
          - maxLength: 64
            minLength: 1
            type: string
          - type: 'null'
          title: Lastname
        phone:
          anyOf:
          - maxLength: 32
            type: string
          - type: 'null'
          title: Phone
        role:
          anyOf:
          - enum:
            - harbormaster
            - boat_owner
            type: string
          - type: 'null'
          title: Role
      title: UserPatch
      type: object
...
  securitySchemes:
    APIKeyCookie:
      in: cookie
      name: dockpulse_access
      type: apiKey
info:
  description: 'REST API for the DockPulse harbor berth availability system.


    The OpenAPI document at `/openapi.json` (and the YAML mirror in

    `docs/api/openapi.yml`) is generated directly from these route

    definitions and Pydantic models. The committed YAML is regenerated by

    `scripts/dump_openapi.py` and CI fails on drift.

    '
  title: DockPulse API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/admin/harbors:
    get:
      operationId: adminListHarbors
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/HarborAdminOut'
                title: Response Adminlistharbors
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Harbors
      tags:
      - admin
    post:
      operationId: adminCreateHarbor
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HarborCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HarborAdminOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Harbor
      tags:
      - admin
  /api/admin/harbors/{harbor_id}:
    delete:
      operationId: adminDeleteHarbor
      parameters:
      - in: path
        name: harbor_id
        required: true
        schema:
          title: Harbor Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Harbor
      tags:
      - admin
    patch:
      operationId: adminPatchHarbor
      parameters:
      - in: path
        name: harbor_id
        required: true
        schema:
          title: Harbor Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HarborPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HarborAdminOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Patch Harbor
      tags:
      - admin
  /api/admin/nodes/{node_id}/decommission:
    post:
      operationId: adminDecommissionNode
      parameters:
      - in: path
        name: node_id
        required: true
        schema:
          title: Node Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecommissionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Decommission Node
      tags:
      - admin
...
  /api/admin/users:
    get:
      operationId: adminListUsers
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserAdminOut'
                title: Response Adminlistusers
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Users
      tags:
      - admin
    post:
      operationId: adminCreateUser
      parameters:
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__routers__admin__users__UserCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreatedOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create User
      tags:
      - admin
  /api/admin/users/{user_id}:
    delete:
      operationId: adminDeleteUser
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete User
      tags:
      - admin
    patch:
      operationId: adminPatchUser
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/app__routers__admin__users__UserPatch'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPatchOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Patch User
      tags:
      - admin
  /api/admin/users/{user_id}/harbor-grants:
    get:
      operationId: adminListUserGrants
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/HarborGrantOut'
                title: Response Adminlistusergrants
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List User Grants
      tags:
      - admin
    post:
      operationId: adminGrantHarbor
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HarborGrant'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrantResultOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Grant Harbor
      tags:
      - admin
  /api/admin/users/{user_id}/harbor-grants/{harbor_id}:
    delete:
      operationId: adminRevokeHarbor
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: path
        name: harbor_id
        required: true
        schema:
          title: Harbor Id
          type: string
      - in: header
        name: Cf-Access-Jwt-Assertion
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cf-Access-Jwt-Assertion
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Revoke Harbor
      tags:
      - admin
...
  /api/harbors:
    get:
      operationId: listHarbors
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/HarborOut'
                title: Response Listharbors
                type: array
          description: Successful Response
      security:
      - APIKeyCookie: []
      summary: List all harbors
      tags:
      - harbors
...
  /api/nodes/{node_id}/decommission:
    post:
      operationId: decommissionNode
      parameters:
      - in: path
        name: node_id
        required: true
        schema:
          title: Node Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeHealthOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Mark node as decommissioned
      tags:
      - nodes
...
servers:
- description: Local development
  url: http://localhost:8000
- description: Staging
  url: https://staging.example.com
...